+++ packages/builder/lib/src/client.dart
@@ -2,7 +2,6 @@ import 'package:either_dart/either.dart';
 import 'package:light_logger/light_logger.dart';
 import 'package:theta/src/data/models/get_page_response.dart';
 import 'package:theta/src/data/models/get_styles_response.dart';
-import 'package:theta/src/domain/usecases/base_usecase.dart';
 import 'package:theta/src/domain/usecases/get_component_usecase.dart';
 import 'package:theta/src/domain/usecases/get_styles_usecase.dart';
 
@@ -24,15 +23,21 @@ class ThetaClient {
         );
       });
 
-  Future<Either<Exception, GetStylesResponseEntity>> _fetchStyles() async =>
-      _getStylesUseCase(Params.empty);
+  Future<Either<Exception, GetStylesResponseEntity>> _fetchStyles(
+          {bool preloadAllowed = true}) async =>
+      _getStylesUseCase(GetStylesUseCaseParams(preloadAllowed: preloadAllowed));
 
   Future<Either<Exception, GetPageResponseEntity>> _fetchComponent(
-          final String componentName) async =>
-      _getComponentUseCase(
-          GetComponentUseCaseParams(componentName: componentName));
+          final String componentName,
+          bool preloadAllowed,
+          String? branchName) async =>
+      _getComponentUseCase(GetComponentUseCaseParams(
+        componentName: componentName,
+        preloadAllowed: preloadAllowed,
+        branchName: branchName,
+      ));
 
-  Future<Either<Exception, GetPageResponseEntity>> build(
-          String componentName) async =>
-      await _fetchComponent(componentName);
+  Future<Either<Exception, GetPageResponseEntity>> build(String componentName,
+          {bool preloadAllowed = true, String? branchName}) async =>
+      await _fetchComponent(componentName, preloadAllowed, branchName);
 }