+++ packages/builder/lib/src/data/datasources/local_styles_service.dart
@@ -1,11 +1,23 @@
 import 'dart:convert';
 
+import 'package:flutter/services.dart';
 import 'package:hive/hive.dart';
+import 'package:light_logger/light_logger.dart';
+import 'package:theta/src/core/constants.dart';
 import 'package:theta/src/data/models/get_styles_response.dart';
+import 'package:theta/src/data/models/preload_file.dart';
+import 'package:theta/src/data/models/token.dart';
 
 class LocalStylesService {
-  const LocalStylesService(this.cacheExtentionInSeconds, this.isCacheEnabled);
+  const LocalStylesService(
+    this._clientToken,
+    this._preloadFile,
+    this.cacheExtentionInSeconds,
+    this.isCacheEnabled,
+  );
 
+  final ClientToken _clientToken;
+  final PreloadFile _preloadFile;
   final int cacheExtentionInSeconds;
   final bool isCacheEnabled;
 
@@ -19,43 +31,51 @@ class LocalStylesService {
     final box = await getBox();
 
     if (box.get('styles') == null) {
+      Logger.printDefault('Cache doesn\'t contain styles');
       return null;
     }
 
     final cachedJson = json.decode(box.get('styles'));
 
     if (cachedJson == null) {
+      Logger.printDefault('Cached styles are null');
       return null;
     }
 
     // if the cached component is older than [cacheExtentionInSeconds] in seconds, return null
-
     final createdAt = cachedJson['created_at'];
     final now = DateTime.now().millisecondsSinceEpoch;
     final diff = now - createdAt;
     if (diff > 1000 * cacheExtentionInSeconds) {
+      Logger.printDefault('Cache expired');
       return null;
     }
 
     // if the cache is not expired, return it
-    if (cachedJson != null) {
-      return GetStylesResponseEntity.fromJson(cachedJson);
-    }
-    return null;
+    return GetStylesResponseEntity.fromJson(cachedJson);
   }
 
-  void saveResponse(GetStylesResponseEntity getStylesResponseEntity) async {
+  Future<void> saveResponse(
+      GetStylesResponseEntity getStylesResponseEntity) async {
     final box = await getBox();
-    box.put(
+    await box.put(
         'styles',
         json.encode({
           ...getStylesResponseEntity.toJson(),
           'created_at': DateTime.now().millisecondsSinceEpoch,
         }));
+    Logger.printDefault('Styles saved in cache');
   }
 
-  void clearCache() async {
+  Future<void> clearCache() async {
     final box = await getBox();
-    box.clear();
+    await box.clear();
+  }
+
+  Future<GetStylesResponseEntity> getPreloadedStyles() async {
+    final res = _preloadFile.customJson ??
+        jsonDecode(await rootBundle.loadString(thetaPreloadFilePath));
+    return GetStylesResponseEntity.fromJson(
+        jsonDecode(decompressAndDecrypt(_clientToken.key, res['styles'])));
   }
 }