+++ packages/builder/lib/src/data/models/preload_flag.dart
@@ -1,5 +1,0 @@
-class PreloadFlag {
-  PreloadFlag(this.value);
-
-  bool value;
-}