+++ packages/builder/lib/src/presentation/local_notifier_provider.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
+import 'package:theta/src/data/models/preload_file.dart';
+import 'package:theta/src/dependency_injection/di.dart';
 import 'package:theta_models/theta_models.dart';
 
 class LocalNotifierProvider extends StatelessWidget {
@@ -17,10 +19,12 @@ class LocalNotifierProvider extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final globalState = context.watch<TreeState>();
+    final isPreloaded = getIt<PreloadFile>().enabled;
     return ChangeNotifierProvider(
       create: (_) => globalState.copyWith(
         workflows: workflows,
         nodeOverrides: nodeOverrides,
+        preloaded: isPreloaded,
       ),
       child: child,
     );