+++ packages/cli/lib/src/domain/repositories/directory_repository.dart
@@ -1,5 +1,10 @@
 import 'package:either_dart/either.dart';
 
 abstract class DirectoryRepository {
-  Future<Either<Exception, void>> writePreloadFile(String key, String json);
+  Future<Either<Exception, void>> writePreloadFile(
+      {required String anonKey,
+      required String jsonKey,
+      required String content});
+
+  Future<Either<Exception, void>> preloadImages(Map<String, dynamic> json);
 }